version 0.6.0

+++ lib/blocs/proxy_list/proxy_list_bloc.dart
@@ -1,7 +1,42 @@
-import 'package:dio/dio.dart';
+import 'package:equatable/equatable.dart';
+import 'package:flibusta/model/connectionCheckResult.dart';
 import 'package:flibusta/services/http_client/http_client.dart';
 import 'package:flibusta/services/local_storage.dart';
+import 'package:flutter/cupertino.dart';
 import 'package:rxdart/rxdart.dart';
+
+@immutable
+class ProxyInfo extends Equatable {
+  final String name;
+  final String hostPort;
+  final bool isDeletable;
+  final BehaviorSubject<ConnectionCheckResult> connectionCheckResultController =
+      BehaviorSubject<ConnectionCheckResult>();
+
+  ProxyInfo(
+    this.hostPort, {
+    this.name,
+    this.isDeletable = false,
+  }) {
+    ProxyHttpClient().connectionCheck(hostPort).then((connectionCheckResult) {
+      connectionCheckResultController.add(connectionCheckResult);
+    });
+  }
+
+  Future<void> checkConnection() async {
+    if (connectionCheckResultController.value == null) return;
+    connectionCheckResultController.add(null);
+    connectionCheckResultController
+        .add(await ProxyHttpClient().connectionCheck(hostPort));
+  }
+
+  Future<dynamic> close() {
+    return connectionCheckResultController.close();
+  }
+
+  @override
+  List<Object> get props => [hostPort, isDeletable];
+}
 
 class ProxyListBloc {
   var _actualProxyController = BehaviorSubject<String>.seeded('');
@@ -10,46 +45,83 @@ class ProxyListBloc {
   //input
   Sink<String> get _actualProxySink => _actualProxyController.sink;
 
-  var _proxyListController = BehaviorSubject<List<String>>.seeded([]);
+  var _proxyListController = BehaviorSubject<List<ProxyInfo>>.seeded([]);
   //output
-  Stream<List<String>> get proxyListStream => _proxyListController.stream;
+  Stream<List<ProxyInfo>> get proxyListStream => _proxyListController.stream;
   //input
-  Sink<List<String>> get _proxyListSink => _proxyListController.sink;
-
-  CancelToken cancelToken;
+  Sink<List<ProxyInfo>> get _proxyListSink => _proxyListController.sink;
 
   ProxyListBloc() {
-    cancelToken = CancelToken();
-    LocalStorage()
-        .getProxies()
-        .then((proxyList) => _proxyListController.add(proxyList));
-    LocalStorage()
-        .getActualProxy()
-        .then((actualProxy) => _actualProxyController.add(actualProxy));
+    LocalStorage().getProxies().then((proxyList) {
+      List<ProxyInfo> initialProxyList = [
+        ProxyInfo(
+          '',
+          name: 'Без прокси',
+        ),
+        ProxyInfo(
+          'flibustauser:ilovebooks@35.217.29.210:1194',
+          name: 'Прокси создателя приложения 1',
+        ),
+        ProxyInfo(
+          'flibustauser:ilovebooks@35.228.73.110:3128',
+          name: 'Прокси создателя приложения 2',
+        ),
+      ];
+      initialProxyList.addAll(
+        proxyList.map((proxyHostPort) {
+          return ProxyInfo(
+            proxyHostPort,
+            isDeletable: true,
+          );
+        }).toList(),
+      );
+      _proxyListController.add(initialProxyList);
+    });
+
+    LocalStorage().getActualProxy().then((actualProxy) {
+      _actualProxyController.add(actualProxy);
+    });
   }
 
-  setActualProxy(String proxy) {
+  void setActualProxy(String proxy) {
     LocalStorage().setActualProxy(proxy);
     ProxyHttpClient().setProxy(proxy);
     _actualProxySink.add(proxy);
   }
 
-  addToProxyList(String proxy) {
+  void checkProxiesConnection() {
+    _proxyListController.value.forEach((proxyInfo) {
+      proxyInfo.checkConnection();
+    });
+  }
+
+  void addToProxyList(String proxy) {
     LocalStorage().addProxy(proxy);
-    var newProxyList = [..._proxyListController.value, proxy];
+    List<ProxyInfo> newProxyList = [
+      ..._proxyListController.value,
+      ProxyInfo(proxy, isDeletable: true),
+    ];
     _proxyListSink.add(newProxyList);
   }
 
-  removeFromProxyList(String proxy) async {
+  void removeFromProxyList(String proxy) async {
+    var currentProxyList = _proxyListController.value;
+    var proxyToDelete = currentProxyList.firstWhere(
+        (proxyInfo) => proxyInfo.hostPort == proxy && proxyInfo.isDeletable);
+
+    if (proxyToDelete == null) return;
+
     await LocalStorage().deleteProxy(proxy);
     if (proxy == await LocalStorage().getActualProxy()) {
-      LocalStorage().setActualProxy('');
+      await LocalStorage().setActualProxy('');
     }
-    _proxyListSink.add([..._proxyListController.value]..remove(proxy));
+    proxyToDelete.close();
+    _proxyListSink.add(currentProxyList..remove(proxyToDelete));
   }
 
   void dispose() {
     _actualProxyController.close();
+    _proxyListController.value.forEach((proxyInfo) => proxyInfo.close());
     _proxyListController.close();
   }
 }
